pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mama-health-ai-challenge"
version = "0.1.0"
description = "AI Data Engineer Challenge - Patient Journey Insights from Reddit"
requires-python = ">=3.11"
authors = [
    {name = "Data Engineer", email = "engineer@mama.health"}
]

dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    
    # Orchestration
    "dagster>=1.7.0",
    "dagster-webserver>=1.7.0",
    "dagster-postgres>=0.21.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Reddit API
    "praw>=7.7.0",
    
    # LLM Interface
    "litellm>=1.40.0",
    "google-generativeai>=0.3.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    
    # Utilities
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mama_health"]

[tool.uv]
# Python version is specified in [project] requires-python = ">=3.11"

[tool.black]
line-length = 100
target-version = ["py311"]
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
